move makeshift sort function to include all filtered document types
The sort function currently applies each document type's native sort function (eg. order by sent date for letters, alphabetically for persons). This is only a provisionally hack and should in the future be enhanced to list exact date matches first and show in the search-results what exactly in the file matched the date search query.

Co-Authored-By: Peter Stadler <2247936+peterstadler@users.noreply.github.com>